I switched persistent storage from Core Data to File Manager

+++ Bookworm/Bookworm/View/UI Components/EmojiRating.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct EmojiRating: View {
-    let rating: Int16
+    let rating: Int
     
     var body: some View {
         switch rating {
